+++ Data/Repositories/Abstract/IDepartmentRepository.cs
@@ -2,7 +2,7 @@
 
 namespace libAPI.Data.Repositories.Abstract
 {
-	public interface IDepartmentRepository:IRepository<Department, libAPIContext,int>
+	public interface IDepartmentRepository:IRepository<Department, libAPIContext, short>
 	{
 	}
 }
