+++ Data/Repositories/Abstract/IEducationalDegreeRepository.cs
@@ -2,7 +2,7 @@
 
 namespace libAPI.Data.Repositories.Abstract
 {
-	public interface IEducationalDegreeRepository: IRepository<EducationalDegree, libAPIContext, int>
+	public interface IEducationalDegreeRepository: IRepository<EducationalDegree, libAPIContext, short>
 	{
 	}
 }
