+++ Data/Repositories/Abstract/ISubCategoryRepository.cs
@@ -2,7 +2,7 @@
 
 namespace libAPI.Data.Repositories.Abstract
 {
-	public interface ISubCategoryRepository:IRepository<SubCategory, libAPIContext,int>
+	public interface ISubCategoryRepository:IRepository<SubCategory, libAPIContext,short>
 	{
 	}
 }
