+++ Data/Repositories/Concrete/EfCore/EfCoreAddressCountryRepository.cs
@@ -4,7 +4,7 @@ using ShopApp.data.Concrete.EfCore;
 
 namespace libAPI.Data.Repositories.Concrete.EfCore
 {
-	public class EfCoreAddressCountryRepository : EfCoreGenericRepository<AddressCountry, libAPIContext,int>,IAddressCountryRepository
+	public class EfCoreAddressCountryRepository : EfCoreGenericRepository<AddressCountry, libAPIContext, short>,IAddressCountryRepository
 	{
 		public EfCoreAddressCountryRepository(libAPIContext context) : base(context)
 		{
