+++ Data/Repositories/Concrete/EfCore/EfCoreEducationalDegreeRepository.cs
@@ -4,7 +4,7 @@ using ShopApp.data.Concrete.EfCore;
 
 namespace libAPI.Data.Repositories.Concrete.EfCore
 {
-	public class EfCoreEducationalDegreeRepository : EfCoreGenericRepository<EducationalDegree, libAPIContext, int>, IEducationalDegreeRepository
+	public class EfCoreEducationalDegreeRepository : EfCoreGenericRepository<EducationalDegree, libAPIContext, short>, IEducationalDegreeRepository
 	{
 		public EfCoreEducationalDegreeRepository(libAPIContext context) : base(context)
 		{
